Add endpoint for RadioBridge LoRa Buttons
- Refactored Flic Buttons into their own file
- Started a routes.js file similar to the one in Sensors
