+++ CHANGELOG.md
@@ -19,6 +19,7 @@ the code was deployed.
 - Send outage notifications to responder phone along with instructions for power-cycling hubs (CU-vf8x83).
 - Send reminders about hub outages (CU-vf8x83).
 - Pairing tool can directly buy phone numbers from Twilio (CU-1tdma20).
+- API endpoint for pressing a Radio Bridge LoRa button (CU-1v85jry).
 
 ### Changed
