+++ server/.env.example
@@ -109,3 +109,11 @@ VITALS_ALERT_THRESHOLD_TEST=900
 # Threshold for sending out a Vitals Alert Reminder to the Responder Phone and Heartbeat Recipients (in seconds)
 SUBSEQUENT_VITALS_ALERT_THRESHOLD=151200
 SUBSEQUENT_VITALS_ALERT_THRESHOLD_TEST=120
+
+# Primary and secondary API keys used by Radio Bridge to send API calls to the Buttons server
+# Two keys are used to allow for seamless API key rotation
+# Get this from Radio Bridge Console --> API --> Header Authorization Code
+RADIO_BRIDGE_API_KEY_PRIMARY=123
+RADIO_BRIDGE_API_KEY_PRIMARY_TEST=123
+RADIO_BRIDGE_API_KEY_SECONDARY=123
+RADIO_BRIDGE_API_KEY_SECONDARY_TEST=123
